+++ chain-signatures/crypto/src/kdf.rs
@@ -4,11 +4,14 @@ use anyhow::Context;
 use k256::{
     ecdsa::{RecoveryId, Signature, VerifyingKey},
     elliptic_curve::{point::AffineCoordinates, sec1::ToEncodedPoint, CurveArithmetic},
-    Scalar, Secp256k1, SecretKey,
+    Scalar, Secp256k1, SecretKey, AffinePoint
 };
-use mpc_primitives::Chain;
+use mpc_primitives::{Chain, Signature as MpcSignature};
 use near_account_id::AccountId;
 use sha3::{Digest, Keccak256, Sha3_256};
+use hkdf::Hkdf;
+// TODO: replace with smaller crate
+use near_primitives::hash::CryptoHash;
 
 // Constant prefix that ensures epsilon derivation values are used specifically for
 // Sig.Network with key derivation protocol vX.Y.Z.
@@ -231,6 +234,89 @@ pub fn recover(
     ))
     .context("Failed to parse returned key")
 }
+
+// TODO: add unit tests
+// In case there are multiple requests in the same block (hence same entropy), we need to ensure
+// that we generate different random scalars as delta tweaks.
+// Receipt ID should be unique inside of a block, so it serves us as the request identifier.
+pub fn derive_delta(
+    request_id: [u8; 32],
+    entropy: [u8; 32],
+    presignature_big_r: AffinePoint,
+) -> Scalar {
+    let hk = Hkdf::<Sha3_256>::new(None, &entropy);
+    let info = format!("{DELTA_DERIVATION_PREFIX}:{}", CryptoHash(request_id));
+    let mut okm = [0u8; 32];
+    hk.expand(info.as_bytes(), &mut okm).unwrap();
+    hk.expand(
+        presignature_big_r.to_encoded_point(true).as_bytes(),
+        &mut okm,
+    )
+    .unwrap();
+    Scalar::from_non_biased(okm)
+}
+
+// Constant prefix that ensures delta derivation values are used specifically for
+// near-mpc-recovery with key derivation protocol vX.Y.Z.
+const DELTA_DERIVATION_PREFIX: &str = "near-mpc-recovery v0.1.0 delta derivation:";
+
+// TODO: add unit tests
+// try to get the correct recovery id for this signature by brute force.
+pub fn into_signature(
+    public_key: &k256::AffinePoint,
+    big_r: &k256::AffinePoint,
+    s: &k256::Scalar,
+    msg_hash: Scalar,
+) -> anyhow::Result<MpcSignature> {
+    let public_key = public_key.to_encoded_point(false);
+    let signature = k256::ecdsa::Signature::from_scalars(x_coordinate(big_r), s)
+        .context("cannot create signature from cait_sith signature")?;
+    let pk0 = recover(
+        &msg_hash.to_bytes()[..],
+        &signature,
+        RecoveryId::try_from(0).context("cannot create recovery_id=0")?,
+    )
+    .context("unable to use 0 as recovery_id to recover public key")?
+    .to_encoded_point(false);
+    if public_key == pk0 {
+        return Ok(MpcSignature::new(*big_r, *s, 0));
+    }
+
+    let pk1 = recover(
+        &msg_hash.to_bytes()[..],
+        &signature,
+        RecoveryId::try_from(1).context("cannot create recovery_id=1")?,
+    )
+    .context("unable to use 1 as recovery_id to recover public key")?
+    .to_encoded_point(false);
+    if public_key == pk1 {
+        return Ok(MpcSignature::new(*big_r, *s, 1));
+    }
+
+    anyhow::bail!("cannot use either recovery id (0 or 1) to recover pubic key")
+}
+
+// TODO: add unit tests
+pub fn valid_signature(root_sk: &k256::SecretKey, args: &mpc_primitives::SignArgs) -> MpcSignature {
+    use k256::ecdsa::signature::hazmat::PrehashSigner as _;
+    use k256::elliptic_curve::point::DecompressPoint as _;
+
+    let derived_secret_key = derive_secret_key(root_sk, args.epsilon);
+    let signing_key = k256::ecdsa::SigningKey::from(&derived_secret_key);
+    let (ecdsa_sig, _): (k256::ecdsa::Signature, _) = signing_key
+        .sign_prehash(&args.payload.to_bytes())
+        .expect("signing should succeed");
+    let (r_bytes, _) = ecdsa_sig.split_bytes();
+    let big_r =
+        k256::AffinePoint::decompress(&r_bytes, k256::elliptic_curve::subtle::Choice::from(0))
+            .unwrap();
+    let s = *ecdsa_sig.s().as_ref();
+    let expected_public_key = derive_key(root_sk.public_key().into(), args.epsilon);
+
+    into_signature(&expected_public_key, &big_r, &s, args.payload)
+        .expect("signature should validate")
+}
+
 
 #[cfg(test)]
 mod tests {
